+++ README.md
@@ -5,6 +5,17 @@ AvroTurf is a library that makes it easier to encode and decode data using the [
 * Provides an idiomatic Ruby interface.
 * Allows referencing schemas defined in another file.
 
+## Deprecation Notice
+
+The `AvroTurf::SchemaRegistry`, `AvroTurf::CachedSchemaRegistry`,
+and `FakeSchemaRegistryServer` names have been deprecated because the Avro spec recently
+introduced an incompatible [single-message encoding format](https://github.com/apache/avro/commit/30408a9c192c5f4eaaf42f01f0ffbfffd705aa57).
+
+These classes have been renamed to `AvroTurf::ConfluentSchemaRegistry`,
+`AvroTurf::CachedConfluentSchemaRegistry`, and `FakeConfluentSchemaRegistry`.
+
+The aliases for the original names will be removed in a future release.
+
 ## Installation
 
 Add this line to your application's Gemfile:
@@ -121,14 +132,14 @@ data = avro.encode({ "title" => "hello, world" }, schema_name: "greeting")
 avro.decode(data) #=> { "title" => "hello, world" }
 ```
 
-### Schema Registry Client
+### Confluent Schema Registry Client
 
-The SchemaRegistry client used by the Messaging API can also be used directly.
+The ConfluentSchemaRegistry client used by the Messaging API can also be used directly.
 It can check whether a schema is compatible with a subject in the registry using the [Compatibility API](http://docs.confluent.io/3.1.2/schema-registry/docs/api.html#compatibility):
 
 ```ruby
 require 'avro_turf'
-require 'avro_turf/schema_registry'
+require 'avro_turf/confluent_schema_registry'
 
 schema = <<-JSON
 {
@@ -147,13 +158,13 @@ schema = <<-JSON
 }
 JSON
 
-registry = AvroTurf::SchemaRegistry.new("http://my-registry:8081/")
+registry = AvroTurf::ConfluentSchemaRegistry.new("http://my-registry:8081/")
 
 # Returns true if the schema is compatible, nil if the subject or version is not registered, and false if incompatible.
 registry.compatible?("person", schema)
 ```
 
-The SchemaRegistry client can also change the global compatibility level or the compatibility level for an individual subject using the [Config API](http://docs.confluent.io/3.1.2/schema-registry/docs/api.html#config):
+The ConfluentSchemaRegistry client can also change the global compatibility level or the compatibility level for an individual subject using the [Config API](http://docs.confluent.io/3.1.2/schema-registry/docs/api.html#config):
 
 ```ruby
 registry.update_global_config(compatibility: 'FULL')
@@ -162,7 +173,7 @@ registry.update_subject_config("person", compatibility: 'NONE')
 
 ### Testing Support
 
-AvroTurf includes a `FakeSchemaRegistryServer` that can be used in tests. The
+AvroTurf includes a `FakeConfluentSchemaRegistryServer` that can be used in tests. The
 fake schema registry server depends on Sinatra but it is _not_ listed as a runtime
 dependency for AvroTurf. Sinatra must be added to your Gemfile or gemspec in order
 to use the fake server.
@@ -170,14 +181,14 @@ to use the fake server.
 Example using RSpec:
 
 ```ruby
-require 'avro_turf/test/fake_schema_registry_server'
+require 'avro_turf/test/fake_confluent_schema_registry_server'
 require 'webmock/rspec'
 
 # within an example
 let(:registry_url) { "http://registry.example.com" }
 before do
-  stub_request(:any, /^#{registry_url}/).to_rack(FakeSchemaRegistryServer)
-  FakeSchemaRegistryServer.clear
+  stub_request(:any, /^#{registry_url}/).to_rack(FakeConfluentSchemaRegistryServer)
+  FakeConfluentSchemaRegistryServer.clear
 end
 
 # Messaging objects created with the same registry_url will now use the fake server.
