+++ avro_turf.gemspec
@@ -27,4 +27,14 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "webmock"
   spec.add_development_dependency "sinatra"
   spec.add_development_dependency "json_spec"
+
+  spec.post_install_message = %{
+avro_turf v0.8.0 deprecates the names AvroTurf::SchemaRegistry,
+AvroTurf::CachedSchemaRegistry, and FakeSchemaRegistryServer.
+
+Use AvroTurf::ConfluentSchemaRegistry, AvroTurf::CachedConfluentSchemaRegistry,
+and FakeConfluentSchemaRegistryServer instead.
+
+See https://github.com/dasch/avro_turf#deprecation-notice
+}
 end
