+++ spec/cached_confluent_schema_registry_spec.rb
@@ -1,9 +1,9 @@
 require 'webmock/rspec'
-require 'avro_turf/cached_schema_registry'
-require 'avro_turf/test/fake_schema_registry_server'
+require 'avro_turf/cached_confluent_schema_registry'
+require 'avro_turf/test/fake_confluent_schema_registry_server'
 
-describe AvroTurf::CachedSchemaRegistry do
-  let(:upstream) { instance_double(AvroTurf::SchemaRegistry) }
+describe AvroTurf::CachedConfluentSchemaRegistry do
+  let(:upstream) { instance_double(AvroTurf::ConfluentSchemaRegistry) }
   let(:registry) { described_class.new(upstream) }
   let(:id) { rand(999) }
   let(:schema) do
@@ -34,8 +34,8 @@ describe AvroTurf::CachedSchemaRegistry do
     end
   end
 
-  it_behaves_like "a schema registry client" do
-    let(:upstream) { AvroTurf::SchemaRegistry.new(registry_url, logger: logger) }
+  it_behaves_like "a confluent schema registry client" do
+    let(:upstream) { AvroTurf::ConfluentSchemaRegistry.new(registry_url, logger: logger) }
     let(:registry) { described_class.new(upstream) }
   end
 end
