--- spec/schema_registry_spec.rb
@@ -1,9 +1,0 @@
-require 'webmock/rspec'
-require 'avro_turf/schema_registry'
-require 'avro_turf/test/fake_schema_registry_server'
-
-describe AvroTurf::SchemaRegistry do
-  it_behaves_like "a schema registry client" do
-    let(:registry) { described_class.new(registry_url, logger: logger) }
-  end
-end
